+++ cassandra-protocol/src/frame/frame_prepare.rs
@@ -4,13 +4,14 @@ use crate::frame::*;
 use crate::types::*;
 
 /// Struct that represents a body of a frame of type `prepare`
-#[derive(Debug)]
+#[derive(Debug, Ord, PartialOrd, Eq, PartialEq, Hash, Clone, Default)]
 pub struct BodyReqPrepare {
     query: String,
 }
 
 impl BodyReqPrepare {
     /// Creates new body of a frame of type `prepare` that prepares query `query`.
+    #[inline]
     pub fn new(query: String) -> BodyReqPrepare {
         BodyReqPrepare { query }
     }
@@ -32,6 +33,13 @@ impl Serialize for BodyReqPrepare {
     }
 }
 
+impl FromCursor for BodyReqPrepare {
+    #[inline]
+    fn from_cursor(cursor: &mut Cursor<&[u8]>) -> error::Result<Self> {
+        from_cursor_str_long(cursor).map(|query| BodyReqPrepare::new(query.into()))
+    }
+}
+
 impl Frame {
     pub fn new_req_prepare(query: String, flags: Flags, version: Version) -> Frame {
         let direction = Direction::Request;
@@ -47,5 +55,22 @@ impl Frame {
             None,
             vec![],
         )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::io::Cursor;
+
+    use crate::frame::frame_prepare::BodyReqPrepare;
+    use crate::frame::FromCursor;
+
+    #[test]
+    fn should_deserialize_body() {
+        let data = [0, 0, 0, 3, 102, 111, 111, 0];
+        let mut cursor = Cursor::new(data.as_slice());
+
+        let body = BodyReqPrepare::from_cursor(&mut cursor).unwrap();
+        assert_eq!(body.query, "foo");
     }
 }