+++ cassandra-protocol/src/frame/frame_ready.rs
@@ -1,13 +1,21 @@
 use std::io::Cursor;
 
-use crate::frame::Serialize;
+use crate::error;
+use crate::frame::{FromCursor, Serialize};
 
-#[derive(Debug, PartialEq, Default)]
+#[derive(Debug, PartialEq, Default, Ord, PartialOrd, Eq, Hash)]
 pub struct BodyResReady;
 
 impl Serialize for BodyResReady {
-    #[inline]
+    #[inline(always)]
     fn serialize(&self, _cursor: &mut Cursor<&mut Vec<u8>>) {}
+}
+
+impl FromCursor for BodyResReady {
+    #[inline(always)]
+    fn from_cursor(_cursor: &mut Cursor<&[u8]>) -> error::Result<Self> {
+        Ok(BodyResReady)
+    }
 }
 
 #[cfg(test)]